refactor: move base_analysis to where it belongs
