+++ freqtrade/optimize/analysis/base_analysis.py
